requirements.txt
selenium
undetected-chromedriver
ultralytics
Pillow
